Move usePrettyJson, date utils, and 2 icons to shared UI
