+++ ui/apps/dev-server-ui/src/components/App/AppCard.tsx
@@ -4,6 +4,7 @@ import { Button } from '@inngest/components/Button';
 import { Link } from '@inngest/components/Link/Link';
 import { IconChevron } from '@inngest/components/icons/Chevron';
 import { IconStatusCircleCheck } from '@inngest/components/icons/StatusCircleCheck';
+import { IconStatusCircleExclamation } from '@inngest/components/icons/StatusCircleExclamation';
 import { IconStatusCircleMinus } from '@inngest/components/icons/StatusCircleMinus';
 import { classNames } from '@inngest/components/utils/classNames';
 import { toast } from 'sonner';
@@ -11,7 +12,7 @@ import { toast } from 'sonner';
 import AppCardHeader from '@/components/App/AppCardHeader';
 import CodeLine from '@/components/Code/CodeLine';
 import useDebounce from '@/hooks/useDebounce';
-import { IconSpinner, IconStatusCircleExclamation } from '@/icons';
+import { IconSpinner } from '@/icons';
 import { useDeleteAppMutation, useUpdateAppMutation, type App } from '@/store/generated';
 import isValidUrl from '@/utils/urlValidation';
 import AppCardStep from './AppCardStep';
