+++ ui/apps/dev-server-ui/src/components/Function/Output.tsx
@@ -1,7 +1,6 @@
 import { CodeBlock } from '@inngest/components/CodeBlock';
+import { usePrettyJson } from '@inngest/components/hooks/usePrettyJson';
 import { renderOutput, type OutputType } from '@inngest/components/utils/outputRenderer';
-
-import { usePrettyJson } from '@/hooks/usePrettyJson';
 
 interface OutputCardProps {
   type: OutputType;
