+++ ui/apps/dev-server-ui/src/components/Timeline/TimelineNode/TimelineNode.tsx
@@ -3,13 +3,13 @@ import { Button } from '@inngest/components/Button';
 import { MetadataGrid } from '@inngest/components/Metadata';
 import { IconChevron } from '@inngest/components/icons/Chevron';
 import { classNames } from '@inngest/components/utils/classNames';
+import { formatMilliseconds } from '@inngest/components/utils/date';
 import { type HistoryNode } from '@inngest/components/utils/historyParser';
 import * as AccordionPrimitive from '@radix-ui/react-accordion';
 import { AnimatePresence, motion } from 'framer-motion';
 
 import TimelineItemHeader from '@/components/AccordionTimeline/TimelineItemHeader';
 import OutputCard from '@/components/Function/Output';
-import { formatMilliseconds } from '@/utils/date';
 import renderTimelineNode from './TimelineNodeRenderer';
 
 type Props = {
